# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HukuPlusCentral API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /retailers:
    get:
      operationId: listRetailers
      tags: [retailers]
      summary: List all retailers
      responses:
        "200":
          description: List of retailers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Retailer"
    post:
      operationId: createRetailer
      tags: [retailers]
      summary: Create a new retailer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRetailerBody"
      responses:
        "201":
          description: Retailer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Retailer"
  /retailers/{retailerId}:
    get:
      operationId: getRetailer
      tags: [retailers]
      summary: Get a retailer by ID
      parameters:
        - name: retailerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Retailer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Retailer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateRetailer
      tags: [retailers]
      summary: Update a retailer
      parameters:
        - name: retailerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRetailerBody"
      responses:
        "200":
          description: Retailer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Retailer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /retailers/{retailerId}/branches:
    get:
      operationId: listBranches
      tags: [retailers]
      summary: List branches for a retailer
      parameters:
        - name: retailerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [retailers]
      summary: Create a branch for a retailer
      parameters:
        - name: retailerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchBody"
      responses:
        "201":
          description: Branch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /retailers/{retailerId}/branches/{branchId}:
    patch:
      operationId: updateBranch
      tags: [retailers]
      summary: Update a branch
      parameters:
        - name: retailerId
          in: path
          required: true
          schema:
            type: integer
        - name: branchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBranchBody"
      responses:
        "200":
          description: Branch updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteBranch
      tags: [retailers]
      summary: Delete a branch
      parameters:
        - name: retailerId
          in: path
          required: true
          schema:
            type: integer
        - name: branchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a team member's role
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamUser"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Retailer:
      type: object
      required: [id, name, contactEmail, contactPhone, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        contactEmail:
          type: [string, 'null']
        contactPhone:
          type: [string, 'null']
        address:
          type: [string, 'null']
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        branchCount:
          type: ["integer", "null"]
    CreateRetailerBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
        contactEmail:
          type: [string, 'null']
        contactPhone:
          type: [string, 'null']
        address:
          type: [string, 'null']
    UpdateRetailerBody:
      type: object
      properties:
        name:
          type: string
        contactEmail:
          type: [string, 'null']
        contactPhone:
          type: [string, 'null']
        address:
          type: [string, 'null']
        isActive:
          type: boolean
    Branch:
      type: object
      required: [id, retailerId, name, isActive, createdAt]
      properties:
        id:
          type: integer
        retailerId:
          type: integer
        name:
          type: string
        location:
          type: [string, 'null']
        contactPhone:
          type: [string, 'null']
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateBranchBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
        location:
          type: [string, 'null']
        contactPhone:
          type: [string, 'null']
    UpdateBranchBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: [string, 'null']
        contactPhone:
          type: [string, 'null']
        isActive:
          type: boolean
...
    TeamUser:
      type: object
      required: [id, email, firstName, lastName, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        role:
          type: string
          enum: [admin, manager, staff]
        createdAt:
          type: string
          format: date-time
    UpdateUserRoleBody:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, manager, staff]