# lib/api-zod/src/generated/types/branch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HukuPlusCentral API
 * OpenAPI spec version: 0.1.0
 */

export interface Branch {
  id: number;
  retailerId: number;
  name: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  isActive: boolean;
  createdAt: Date;
}