# lib/api-zod/src/generated/types/retailer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HukuPlusCentral API
 * OpenAPI spec version: 0.1.0
 */

export interface Retailer {
  id: number;
  name: string;
  /** @nullable */
  contactEmail: string | null;
  /** @nullable */
  contactPhone: string | null;
  /** @nullable */
  address?: string | null;
  isActive: boolean;
  createdAt: Date;
  /** @nullable */
  branchCount?: number | null;
}