# lib/api-zod/src/generated/types/updateRetailerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HukuPlusCentral API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateRetailerBody {
  name?: string;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  address?: string | null;
  isActive?: boolean;
}